requirements.txt
Django
djangorestframework
markdown
django-filter
pymysql
redis
mysqlclient==1.4.4
